+++ example/lib/main.dart
@@ -59,32 +59,26 @@ class _MyAppState extends State<MyApp> {
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
                 ContainerNeo(
+                  offset: 3,
                   animate: true,
                   padding: const EdgeInsets.all(10),
-                  defaultPaddingBottom: 5,
-                  defaultPaddingRight: 5,
                   bgcolor: const Color.fromARGB(255, 251, 209, 70),
                   borderRadius: 3,
                   onPressed: () {
                     debugPrint('Container_Neo pressed');
                   },
-                  onLongPressed: () {
-                    debugPrint('Container_Neo long pressed');
-                  },
                   child: const Text('Neo Container',
                       style: TextStyle(fontSize: 20)),
                 ),
                 const SizedBox(height: 20),
-                InputField(
-                  padding: const EdgeInsets.all(6),
+                const InputField(
+                  keyboardType: TextInputType.emailAddress,
+                  hintText: "Hint",
+                  hintStyle: TextStyle(color: Colors.blueAccent),
+                  textStyle: TextStyle(color: Colors.amber),
                   width: 250,
-                  hintText: 'Please enter your name',
                   obscureText: false,
-                  controller: null,
-                  bgcolor: const Color.fromARGB(255, 220, 215, 215),
-                  onChanged: () {
-                    debugPrint('InputField changed');
-                  },
+                  bgcolor: Color.fromARGB(255, 220, 215, 215),
                 ),
                 const SizedBox(height: 20),
                 const TextContainer(
@@ -94,19 +88,21 @@ class _MyAppState extends State<MyApp> {
                     width: 150,
                     bgcolor: Color.fromARGB(255, 58, 193, 255)),
                 const SizedBox(height: 20),
-                NeoCircleIconButon(
-                  padding: const EdgeInsets.all(6),
-                  onPressed: () => debugPrint("Button pressed!"),
-                  icon: const Icon(
-                    Icons.add,
-                    size: 60,
-                    color: Colors.black,
-                  ),
-                  backgroundColor: const Color.fromARGB(255, 58, 193, 255),
-                  offsetA: 4,
-                  offsetB: 4,
+                NeoCircleIconButton(
+                    padding: const EdgeInsets.all(6),
+                    icon: const Icon(
+                      Icons.add,
+                      size: 60,
+                      color: Colors.black,
+                    ),
+                    backgroundColor: const Color.fromARGB(255, 58, 193, 255),
+                    offset: 4,
+                    onPressed: () {
+                      debugPrint("Button pressed");
+                    }),
+                const SizedBox(
+                  height: 20,
                 ),
-                const SizedBox(height: 20),
                 NeoSwitch(
                   onChanged: _onSwitchChanged,
                   defaultState: true,