+++ lib/src/widgets/input_field.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 
-// ignore: must_be_immutable
 class InputField extends StatefulWidget {
-  InputField({
+  const InputField({
     required this.hintText,
     required this.width,
     required this.obscureText,
@@ -10,47 +9,73 @@ class InputField extends StatefulWidget {
     required this.bgcolor,
     this.onChanged,
     this.padding,
+    this.keyboardType,
+    this.textStyle,
+    this.hintStyle,
     super.key,
   });
 
-  Color bgcolor;
+  final Color bgcolor;
   final EdgeInsets? padding;
   final double width;
   final String hintText;
-  TextEditingController? controller;
+  final TextEditingController? controller;
   final bool obscureText;
-  Function? onChanged;
+  final Function(String)? onChanged;
+  final TextInputType? keyboardType;
+  final TextStyle? textStyle;
+  final TextStyle? hintStyle;
 
   @override
   State<InputField> createState() => _InputFieldState();
 }
 
 class _InputFieldState extends State<InputField> {
+  late Color currentBgColor;
+
+  @override
+  void initState() {
+    super.initState();
+    currentBgColor = widget.bgcolor;
+  }
+
   @override
   Widget build(BuildContext context) {
     return Container(
       padding: widget.padding ?? const EdgeInsets.all(6),
       decoration: BoxDecoration(
+        color: currentBgColor,
         border: Border.all(
-            color: const Color.fromARGB(255, 0, 0, 0),
-            width: widget.width / 100),
-        boxShadow: [
-          BoxShadow(
-            color: widget.bgcolor,
-          ),
-        ],
+          color: const Color.fromARGB(255, 0, 0, 0),
+          width: widget.width / 100,
+        ),
       ),
       width: widget.width,
       child: TextField(
         onChanged: (String value) {
-          setState(() {
-            widget.bgcolor = Colors.white;
-          });
+          if (currentBgColor != Colors.white) {
+            setState(() {
+              currentBgColor = Colors.white;
+            });
+          }
+          widget.onChanged?.call(value);
         },
         controller: widget.controller,
         obscureText: widget.obscureText,
+        autocorrect: !widget.obscureText,
+        enableSuggestions: !widget.obscureText,
+        keyboardType: widget.keyboardType,
+
+        // Apply the input text style
+        style: widget.textStyle ??
+            const TextStyle(color: Colors.black, fontSize: 16),
+
         decoration: InputDecoration(
-            hintText: widget.hintText, border: InputBorder.none),
+          hintText: widget.hintText,
+          // Apply the hint text style
+          hintStyle: widget.hintStyle ?? const TextStyle(color: Colors.grey),
+          border: InputBorder.none,
+        ),
       ),
     );
   }