+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0(27-12-2025)
+- Updated the gradle version of the example project.
+- ContainerNeo Press animation causing page layout to shift has been fixed. Animation logic overhauled and is no longer gets stuck on pressed state.
+- InputField: Autocorrect and suggestions are disabled when the text is obscured. Added an option to change the keyboard type and customize the text styles of hint and input texts.
+- NeoCircleIconButton: Added a pressing down animation.
+
 ## 0.1.0(28-02-2025)
 - Added 3 new widgets `NeoCircleIconButton` , `NeoSegmentedOptions` and `NeoSwitch`
 - Integrated better performance optimizations and bug fixes.